fastapi_auth_service/requirements.txt
fastapi==0.115.1
uvicorn==0.34.0
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
PyJWT==2.10.1
python-decouple==3.8